Add files via upload
The Random Forest algorithm is a versatile machine learning technique used for both classification and regression tasks. It works by constructing multiple decision trees during training and combining their outputs to improve accuracy and reduce overfitting. Each tree in the forest is trained on a random subset of the data, and features, making the model robust and less sensitive to noise. Predictions are made by aggregating the results from all the trees, either through majority voting for classification or averaging for regression. Random Forest is popular for its simplicity, flexibility, and effectiveness in handling large datasets with high dimensionality.